Bug: sendReminderMessage broke in checkerHandler
