+++ functions/definitions/common/sendFactCheckerMessages.js
@@ -1,3 +1,4 @@
+const admin = require("firebase-admin");
 const { sendWhatsappTextListMessage, sendWhatsappTextMessage, sendWhatsappButtonMessage } = require("./sendWhatsappMessage");
 const { getResponsesObj } = require("./responseUtils");
 
@@ -112,7 +113,7 @@ exports.sendVotingMessage = async function sendVotingMessage(voteRequestSnap, me
   }
 };
 
-exports.sendReminderMessage = async function (to, numOutstanding, voteRequestPath) {
+async function _sendReminderMessage(to, numOutstanding, voteRequestPath) {
   const responses = await getResponsesObj("factChecker");
   const buttons = [{
     type: "reply",
@@ -123,3 +124,33 @@ exports.sendReminderMessage = async function (to, numOutstanding, voteRequestPat
   }];
   await sendWhatsappButtonMessage("factChecker", to, responses.OUTSTANDING_REMINDER.replace("{{num_outstanding}}", `${numOutstanding}`), buttons);
 }
+
+exports.sendRemainingReminder = async function (factCheckerId, platform) {
+  const db = admin.firestore();
+  try {
+    const outstandingVoteRequestsQuerySnap = await db.collectionGroup('voteRequests').where('platformId', '==', factCheckerId).where("category", "==", null).get();
+    const remainingCount = outstandingVoteRequestsQuerySnap.size;
+    if (remainingCount == 0) {
+      await sendWhatsappTextMessage("factChecker", factCheckerId, "Great, you have no further messages to assess. Keep it up!💪");
+      return;
+    }
+    const unassessedMessagesQuerySnap = await db.collection("messages").where("isAssessed", "==", false).get();
+    const unassessedMessageIdList = unassessedMessagesQuerySnap.docs.map((docSnap) => docSnap.id);
+    //sort outstandingVoteRequestsQuerySnap by whether the parent message is assessed
+    const sortedVoteRequestDocs = outstandingVoteRequestsQuerySnap.docs.sort((a, b) => {
+      const aIsAssessed = unassessedMessageIdList.includes(a.ref.parent.parent.id);
+      const bIsAssessed = unassessedMessageIdList.includes(b.ref.parent.parent.id);
+      if (aIsAssessed && !bIsAssessed) {
+        return -1;
+      }
+      if (!aIsAssessed && bIsAssessed) {
+        return 1;
+      }
+      return 0;
+    });
+    const nextVoteRequestPath = sortedVoteRequestDocs[0].ref.path;
+    await _sendReminderMessage(factCheckerId, remainingCount, nextVoteRequestPath);
+  } catch (error) {
+    functions.logger.error("Error sending remaining reminder", error);
+  }
+};
